requirements.txt
# List your required libraries here 
python-pptx
streamlit
pandas
elasticsearch
PyYAML
fpdf
dataframe_image
matplotlib